+++ record_macos/macos/record_macos/Sources/record_macos/RecordConfig.swift
@@ -12,11 +12,17 @@ public enum AudioEncoder: String {
   case wav = "wav"
 }
 
+public enum PcmFormat: String {
+  case int16 = "int16"
+  case float32 = "float32"
+}
+
 public class RecordConfig {
   let encoder: String
   var bitRate: Int
   var sampleRate: Int
   var numChannels: Int
+  let pcmFormat: PcmFormat
   let device: Device?
   let autoGain: Bool
   let echoCancel: Bool
@@ -28,13 +34,15 @@ public class RecordConfig {
   var isModified: Bool {
     bitRate    != (m_args["bitRate"]      as? Int ?? 128000) ||
     sampleRate != (m_args["sampleRate"]   as? Int ?? 44100)  ||
-    numChannels != (m_args["numChannels"] as? Int ?? 2)
+    numChannels != (m_args["numChannels"] as? Int ?? 2) ||
+    pcmFormat.rawValue != (m_args["pcmFormat"] as? String ?? PcmFormat.int16.rawValue)
   }
 
   init(encoder: String,
        bitRate: Int,
        sampleRate: Int,
        numChannels: Int,
+       pcmFormat: PcmFormat = .int16,
        device: Device? = nil,
        autoGain: Bool = false,
        echoCancel: Bool = false,
@@ -45,6 +53,7 @@ public class RecordConfig {
     self.bitRate = bitRate
     self.sampleRate = sampleRate
     self.numChannels = numChannels
+    self.pcmFormat = pcmFormat
     self.device = device
     self.autoGain = autoGain
     self.echoCancel = echoCancel
@@ -86,6 +95,7 @@ extension RecordConfig {
     map["bitRate"] = bitRate
     map["sampleRate"] = sampleRate
     map["numChannels"] = numChannels
+    map["pcmFormat"] = pcmFormat.rawValue
     return map
   }
 
@@ -99,6 +109,7 @@ extension RecordConfig {
       bitRate: args["bitRate"] as? Int ?? 128000,
       sampleRate: args["sampleRate"] as? Int ?? 44100,
       numChannels: args["numChannels"] as? Int ?? 2,
+      pcmFormat: (args["pcmFormat"] as? String).flatMap(PcmFormat.init(rawValue:)) ?? .int16,
       device: device,
       autoGain: args["autoGain"] as? Bool ?? false,
       echoCancel: args["echoCancel"] as? Bool ?? false,