+++ record/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 7.1.1
-- feat: Add configurable 16-bit integer and 32-bit float PCM output on Android and iOS.
+- feat: Add configurable 16-bit integer and 32-bit float PCM output on Android, iOS, macOS, and Windows.
 - fix(ios): Guard ios specific calls to avoid "Recorder has not yet been created..." exception.
 
 ## 7.1.0